- Moved old tests and old data to legacy packages.  - Added JUnit 5
