+++ src/test/java/com/flipkart/zjsonpatch/legacy/AddOperationTest.java
@@ -14,29 +14,16 @@
  * limitations under the License.
 */
 
-package com.flipkart.zjsonpatch;
-
-import org.junit.runners.Parameterized;
+package com.flipkart.zjsonpatch.legacy;
 
 import java.io.IOException;
 import java.util.Collection;
+import org.junit.runners.Parameterized.Parameters;
 
-/**
- * @author ctranxuan (streamdata.io).
- *
- * These tests comes from JS JSON-Patch libraries (
- * https://github.com/Starcounter-Jack/JSON-Patch/blob/master/test/spec/json-patch-tests/tests.json
- * https://github.com/cujojs/jiff/blob/master/test/json-patch-tests/tests.json)
- */
-public class JsLibSamplesTest extends AbstractTest {
+public class AddOperationTest extends AbstractTest {
 
-    @Parameterized.Parameters
+    @Parameters
     public static Collection<PatchTestCase> data() throws IOException {
-        return PatchTestCase.load("js-libs-samples");
-    }
-
-    @Override
-    protected boolean matchOnErrors() {
-        return false;
+        return PatchTestCase.load("add");
     }
 }
