+++ src/test/java/com/flipkart/zjsonpatch/legacy/JsLibSamplesTest.java
@@ -14,16 +14,29 @@
  * limitations under the License.
 */
 
-package com.flipkart.zjsonpatch;
+package com.flipkart.zjsonpatch.legacy;
+
+import org.junit.runners.Parameterized;
 
 import java.io.IOException;
 import java.util.Collection;
-import org.junit.runners.Parameterized.Parameters;
 
-public class AddOperationTest extends AbstractTest {
+/**
+ * @author ctranxuan (streamdata.io).
+ *
+ * These tests comes from JS JSON-Patch libraries (
+ * https://github.com/Starcounter-Jack/JSON-Patch/blob/master/test/spec/json-patch-tests/tests.json
+ * https://github.com/cujojs/jiff/blob/master/test/json-patch-tests/tests.json)
+ */
+public class JsLibSamplesTest extends AbstractTest {
 
-    @Parameters
+    @Parameterized.Parameters
     public static Collection<PatchTestCase> data() throws IOException {
-        return PatchTestCase.load("add");
+        return PatchTestCase.load("js-libs-samples");
+    }
+
+    @Override
+    protected boolean matchOnErrors() {
+        return false;
     }
 }
