+++ src/test/java/com/flipkart/zjsonpatch/legacy/MoveOperationTest.java
@@ -14,10 +14,11 @@
  * limitations under the License.
 */
 
-package com.flipkart.zjsonpatch;
+package com.flipkart.zjsonpatch.legacy;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.flipkart.zjsonpatch.JsonDiff;
 import org.junit.Test;
 import org.junit.runners.Parameterized;
