+++ src/test/java/com/flipkart/zjsonpatch/legacy/RemoveOperationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
 */
 
-package com.flipkart.zjsonpatch;
+package com.flipkart.zjsonpatch.legacy;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -23,10 +23,10 @@ import org.junit.runners.Parameterized;
 /**
  * @author ctranxuan (streamdata.io).
  */
-public class ReplaceOperationTest extends AbstractTest {
+public class RemoveOperationTest extends AbstractTest {
 
     @Parameterized.Parameters
     public static Collection<PatchTestCase> data() throws IOException {
-        return PatchTestCase.load("replace");
+        return PatchTestCase.load("remove");
     }
 }
