+++ src/test/java/com/flipkart/zjsonpatch/legacy/Rfc6902SamplesTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
 */
 
-package com.flipkart.zjsonpatch;
+package com.flipkart.zjsonpatch.legacy;
 
 import org.junit.runners.Parameterized;
