+++ src/test/resources/legacy-data/js-libs-samples-unsupported.json
