+++ src/test/resources/legacy-data/rfc6902-samples-unsupported.json
